Removed cache tag invalidation from importQueueWorker, add required services to factory classes, adjust kernel tests
